Create perceptron_classifier.py
This simple code in python creates a perceptron classifier using the Iris dataset from sklearn. 

The perceptron takes two inputs (the petal length and the petal width) and two types of labels (True: for "Iris Setosa" and False: for "No Iris setosa"). 

For the prediction step you only need to add new data and the model will classify the inputs as "True" = It is an Iris setosa flower or "False" =  It is not an Iris setosa flower.